// src/app/models/nummacdos.model.ts
export interface Cuenta{
    uid: string;
    nombre: string;
    costo: string;
    horariodias: string;
    numreservasmes: string;
    numreservasano: string;
    letra: string;
    numero: string;
    generarSaldoReservar: string;
    puedereservar: string;
    puederegistrarreserva: string;
    debeirpovisita: string;
    mostrarlistamorosos: string;
    generarfoliopago: string;
    
}